moving to a typescript basis and removing the gen AI element for inital release
