+++ docker-compose.yaml
@@ -30,7 +30,17 @@ services:
     volumes:
       - files:/backend/uploads
 
+  nexus:
+    container_name: nexus
+    image: sonatype/nexus3
+    ports:
+      - 8081:8081
+    volumes:
+      - nexus-data:/nexus-data
+    restart: unless-stopped
+
 volumes:
   mongo: {}
   ollama: {}
   files: {}
+  nexus-data: {}
